ss58_service/ss58/requirements.ss58.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.1
requests>=2.32.0
PyNaCl>=1.5.0
